:recycle: Extract collaborators from Plugin.php (792 → 405 lines)
Pure refactor splitting Plugin.php into four single-purpose classes that
sit next to their domain peers, leaving Plugin as the orchestrator.

- src/Util/StubFileFinder: filesystem scan + version directory selection
  for stub registration. Existing PluginVersionStubsTest moves to
  tests/Unit/Util/StubFileFinderTest with #[CoversClass] retargeted.
- src/Handlers/Eloquent/Schema/MigrationSchemaBuilder: discover, parse,
  cache, and report on SQL dumps + PHP migrations. Constructor takes
  Application, Codebase, MigrationCache so the dependencies are explicit
  rather than reaching for ProjectAnalyzer::getInstance() inside.
- src/Providers/AliasStubProvider: writes aliases.phpstub and registers
  it in one call, mirroring CarbonStubProvider::register shape (the
  previous extraction split write and registration across __invoke and
  registerStubs).
- src/Util/InternalErrorReporter: top-level catch handler with the same
  best-effort URL fallback and failOnInternalError rethrow behaviour.

The databasePath() guard remains in Plugin::buildSchema so the builder's
contract stays "always returns a SchemaAggregator." Boot-path only;
reviewed for behaviour parity, allocation neutrality, error preservation,
and Laravel 12-13 / Psalm 7 API correctness.
